pyproject.toml
[project]
name = "stock-agent-ops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "comet-ml>=3.52.1",
    "dagshub>=0.6.3",
    "fastapi>=0.116.1",
    "feast>=0.20.0",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.4",
    "matplotlib>=3.10.6",
    "mlflow>=3.3.2",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "scikit-learn>=1.7.2",
    "torch>=2.8.0",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
